+++ __tests__/mock/mock-types.js
@@ -12,6 +12,8 @@ export function mockUser(fields = {}) {
 		group_ids_names: [],
 		oauth_accounts: [],
 		profile_id: 1,
+		project_dir: '/tmp',
+		slurm_accounts: [],
 		...fields
 	};
 }