+++ src/routes/+page.svelte
@@ -5,17 +5,37 @@
 	import { formatMarkdown } from '$lib/common/component_utilities';
 	import { onMount } from 'svelte';
 
-	let userLoggedIn = $derived(!!page.data.userInfo);
+	/** @type {import('fractal-components/types/api').User|undefined} */
+	let userInfo = $derived(page.data.userInfo);
 	let news = $derived(page.data.news);
 
 	let mounted = $state(false);
 
+	let temporaryMessage = $state('');
+
 	onMount(() => {
 		mounted = true;
+		if (sessionStorage && sessionStorage.getItem('temporary-message')) {
+			temporaryMessage = sessionStorage.getItem('temporary-message') || '';
+			sessionStorage.removeItem('temporary-message');
+		}
 	});
 </script>
 
 <div class="container mt-3">
+	{#if userInfo && !userInfo.is_superuser && (userInfo.profile_id === null || !userInfo.is_verified)}
+		<div class="alert alert-warning">
+			<i class="bi bi-exclamation-triangle"></i>
+			This user is not authorized to use this Fractal instance - please contact {env.PUBLIC_FRACTAL_ADMIN_SUPPORT_EMAIL}.
+		</div>
+	{/if}
+
+	{#if temporaryMessage}
+		<div class="alert alert-info">
+			{temporaryMessage}
+		</div>
+	{/if}
+
 	<img alt="Fractal logo" src={logoMedium} class="float-end" id="fractal-logo-home" />
 
 	<h1 class="fw-light">Welcome to Fractal web client.</h1>
@@ -37,7 +57,7 @@
 		</p>
 
 		<div class="col mb-3">
-			{#if !userLoggedIn}
+			{#if !userInfo}
 				<a href="/auth/login" class="btn btn-primary">Login</a>
 			{/if}
 			<a href="/v2/projects" class="btn btn-primary">Projects</a>