+++ tests/v2/images.spec.js
@@ -13,6 +13,7 @@ test('Dataset images [v2]', async ({ page, project }) => {
 		await createDatasetButton.click();
 		await modal.waitFor();
 		await modal.getByRole('textbox', { name: 'Dataset Name' }).fill('test-dataset');
+		await modal.getByRole('button', { name: 'Advanced options' }).click();
 		await modal.getByRole('textbox', { name: 'Zarr dir' }).fill(randomPath);
 		await modal.getByRole('button', { name: 'Save' }).click();
 		await waitModalClosed(page);